+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log for lambda-stash
 
+## 2.0.0 [2018/05/23]
+
+- Adding support for ELB Classic Load Balancer (elbv1) and Application Load
+Balancer (elbv2) log formats.
+- Adding Elasticsearch requestTimeout configuration.
+- Fixing http-aws-es dependency.
+
 ## 1.1.0 [2017/01/20]
 
 - Allowing custom handlers to be provided through the configuration script.